requirements.txt
anthropic==0.34.1
fitz==0.0.1.dev2
openai==1.42.0
pdf2image==1.17.0
Pillow==9.5.0
Pillow==10.4.0
